moved isShopOpen and its state to its own file to fix issue with modifier importing from util
