+++ game/scripts/vscripts/util.ts
@@ -276,16 +276,3 @@ export function removeContextEntityIfExists(context: TutorialContext, entityKey:
 export function unitIsValidAndAlive(unit: CDOTA_BaseNPC | undefined): boolean {
     return unit !== undefined && IsValidEntity(unit) && unit.IsAlive()
 }
-
-let shopOpen = false;
-CustomGameEventManager.RegisterListener("shop_open_changed", (source, event) => {
-    shopOpen = event.open !== 0;
-})
-
-/**
- * Returns whether the shop ui is currently open.
- * @returns Whether the shop ui is currently open.
- */
-export function isShopOpen() {
-    return shopOpen;
-}
